requirements.txt
# Base packages:
# Flask
# gunicorn

Flask==2.0.3
gunicorn==20.1.0      #19.7.1
itsdangerous==2.0
Jinja2==3.1.0
Werkzeug==2.0
requests==2.27.0
Pillow==8.2.0
opencv-python==4.5.5.64
numpy==1.22.3
firebase-admin==5.2.0
ml_collections==0.1.1
tensorflow-addons
pycocotools
tensorflow-datasets
opencv-python

# Flask==0.12.2
# gunicorn==19.7.1
# itsdangerous==0.24
# Jinja2==2.10
# Werkzeug==0.13
# requests==2.27.0